Fix #517

+++ colplus-ws/src/main/java/org/col/assembly/SectorDelete.java
@@ -12,6 +12,7 @@ import org.col.api.model.ColUser;
 import org.col.api.model.Sector;
 import org.col.api.model.SectorImport;
 import org.col.api.vocab.Datasets;
+import org.col.dao.SectorDao;
 import org.col.db.mapper.NameUsageMapper;
 import org.col.db.mapper.SectorImportMapper;
 import org.col.db.mapper.SectorMapper;
@@ -76,11 +77,18 @@ public class SectorDelete extends SectorRunnable {
     if (!visitedSectors.contains(sectorKey)) {
       visitedSectors.add(sectorKey);
       try (SqlSession session = factory.openSession(true)) {
+        Sector s = session.getMapper(SectorMapper.class).get(sectorKey);
+        if (s == null) {
+          throw new IllegalArgumentException("Sector "+sectorKey+" does not exist");
+        }
         NameUsageMapper um = session.getMapper(NameUsageMapper.class);
         int count = um.deleteBySector(catalogueKey, sectorKey);
         String sectorType = sectorKey == this.sectorKey ? "sector" : "subsector";
         LOG.info("Deleted {} existing taxa with their synonyms and related information from {} {}", count, sectorType, sectorKey);
       
+        // update datasetSectors counts
+        SectorDao.incSectorCounts(session, s, -1);
+        
         session.getMapper(SectorImportMapper.class).delete(sectorKey);
         session.getMapper(SectorMapper.class).delete(sectorKey);
         LOG.info("Deleted {} {}", sectorType, sectorKey);
